requirements.txt
absl-py==1.3.0
appdirs==1.4.4
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///C:/ci/argon2-cffi-bindings_1644551690056/work
arviz @ file:///tmp/build/80754af9/arviz_1614019183254/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///C:/ci/beautifulsoup4_1650293025093/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
blurhash==1.1.4
boost==0.1
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
cachetools==5.2.0
certifi==2022.9.24
cffi @ file:///C:/Windows/Temp/abs_6808y9x40v/croots/recipe/cffi_1659598653989/work
cftime @ file:///C:/ci/cftime_1638358653367/work
charset-normalizer==2.1.1
cmake==3.25.0
cmdstanpy==1.0.8
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
convertdate @ file:///home/conda/feedstock_root/build_artifacts/convertdate_1642883757836/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython @ file:///C:/b/abs_0438epndxm/croots/recipe/cython_1663692771227/work
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dlib @ file:///C:/Users/s2689/dlib-19.22.99-cp39-cp39-win_amd64.whl
dnspython==2.2.1
entrypoints @ file:///C:/ci/entrypoints_1649926621128/work
ephem @ file:///D:/bld/ephem_1649637375602/work
et-xmlfile==1.1.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_ebruxzvd08/croots/recipe/python-fastjsonschema_1661376484940/work
flatbuffers==22.12.6
fonttools==4.25.0
gast==0.4.0
gh==0.0.4
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
h5py==3.7.0
hijri-converter @ file:///home/conda/feedstock_root/build_artifacts/hijri-converter_1653348285237/work
holidays @ file:///home/conda/feedstock_root/build_artifacts/holidays_1669298368598/work
idna==3.4
imageio==2.22.4
importlib-metadata==5.1.0
ipykernel @ file:///C:/b/abs_21ykzkm7y_/croots/recipe/ipykernel_1662361803478/work
ipython @ file:///C:/ci/ipython_1657634415474/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib==1.2.0
jsonschema @ file:///C:/b/abs_59eyhnbyej/croots/recipe/jsonschema_1663375476535/work
jupyter_client @ file:///C:/b/abs_8fbm7986b_/croots/recipe/jupyter_client_1662504374117/work
jupyter_core @ file:///C:/b/abs_a9330r1z_i/croots/recipe/jupyter_core_1664917313457/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
keras==2.11.0
kiwisolver @ file:///C:/ci/kiwisolver_1653292407425/work
korean-lunar-calendar @ file:///home/conda/feedstock_root/build_artifacts/korean_lunar_calendar_1663341251025/work
libclang==14.0.6
libpython==0.2
LunarCalendar==0.0.9
lxml==4.9.1
Markdown==3.4.1
MarkupSafe @ file:///C:/ci/markupsafe_1654508077284/work
Mastodon.py==1.7.0
matplotlib @ file:///C:/ci/matplotlib-suite_1660169687702/work
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
mistune @ file:///C:/ci/mistune_1607359457024/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
multitasking==0.0.11
munkres==1.1.4
nbclient @ file:///C:/ci/nbclient_1650290387259/work
nbconvert @ file:///C:/ci/nbconvert_1649741016669/work
nbformat @ file:///C:/b/abs_1dw90o2uqb/croots/recipe/nbformat_1663744957967/work
nest-asyncio @ file:///C:/ci/nest-asyncio_1649829929390/work
netCDF4 @ file:///C:/ci/netcdf4_1643100823749/work
networkx==2.8.8
notebook @ file:///C:/Windows/TEMP/abs_79abr1_60s/croots/recipe/notebook_1659083661851/work
numexpr @ file:///C:/Windows/Temp/abs_e2036a32-9fe9-47f3-a04c-dbb1c232ba4b334exiur/croots/recipe/numexpr_1656940304835/work
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/b/abs_cdcgk91igc/croots/recipe/pandas_1663772960432/work
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
patsy==0.5.3
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
prometheus-client @ file:///C:/Windows/TEMP/abs_ab9nx8qb08/croots/recipe/prometheus_client_1659455104602/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
prophet==1.1.1
protobuf==3.19.6
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyMeeus==0.5.11
pymongo==4.3.2
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
pyrsistent @ file:///C:/ci/pyrsistent_1636093225342/work
pystan==2.19.1.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-magic==0.4.27
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
PyWavelets==1.4.1
pywin32==302
pywinpty @ file:///C:/ci_310/pywinpty_1644230983541/work/target/wheels/pywinpty-2.0.2-cp39-none-win_amd64.whl
pyzmq @ file:///C:/ci/pyzmq_1657615952984/work
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
setuptools-git==1.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smmap==5.0.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
SQLAlchemy==1.4.44
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
statsmodels==0.13.5
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminado @ file:///C:/ci/terminado_1644322780199/work
testpath @ file:///C:/Windows/TEMP/abs_23c7fa33-cbb9-46dc-b7c5-590c38e2de3d4bmbngal/croots/recipe/testpath_1655908553202/work
threadpoolctl==3.1.0
tifffile==2022.10.10
tornado @ file:///C:/ci/tornado_1662458743919/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3==1.26.12
wbdata==0.3.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
xarray @ file:///opt/conda/conda-bld/xarray_1639166117697/work
yfinance==0.1.86
zipp==3.11.0